refactor: webapi - move /chains endpoints to the chain controller
