+++ clients/chainclient/uploadblob.go
@@ -26,7 +26,7 @@ func (c *Client) UploadBlob(ctx context.Context, fields dict.Dict) (hashing.Hash
 		return hashing.NilHash, nil, nil, err
 	}
 
-	receipt, _, err := c.WaspClient.RequestsApi.WaitForRequest(ctx, c.ChainID.String(), req.ID().String()).Execute()
+	receipt, _, err := c.WaspClient.ChainsApi.WaitForRequest(ctx, c.ChainID.String(), req.ID().String()).Execute()
 
 	return blobHash, req, receipt, err
 }
