+++ packages/webapi/controllers/chain/callview.go
@@ -1,4 +1,4 @@
-package requests
+package chain
 
 import (
 	"net/http"
