+++ packages/webapi/controllers/chain/receipt.go
@@ -1,4 +1,4 @@
-package requests
+package chain
 
 import (
 	"errors"
